requirements.txt
pandacan==0.0.9
prompt-toolkit==3.0.5
